organizing folders in results && prettier
