+++ packages/app/cypress/e2e/page-objects/components/NavbarComponent.ts
@@ -1,13 +1,9 @@
 export class NavbarComponent {
   race(): Cypress.Chainable {
-    return cy.get(
-      "[data-cy='Race-main-nav-link']"
-    );
+    return cy.get("[data-cy='Race-main-nav-link']");
   }
 
   leaderboard(): Cypress.Chainable {
-    return cy.get(
-      "[data-cy='Leaderboard-main-nav-link']"
-    );
+    return cy.get("[data-cy='Leaderboard-main-nav-link']");
   }
 }
