+++ packages/app/cypress/e2e/page-objects/pages/RacePage.ts
@@ -1,6 +1,8 @@
 export class RacePage {
   practiceCardLanguageDropdown(): Cypress.Chainable {
-    return cy.get("[data-cy='practice-card'] [data-cy='language-dropdown'] [data-cy='search-language-input']");
+    return cy.get(
+      "[data-cy='practice-card'] [data-cy='language-dropdown'] [data-cy='search-language-input']"
+    );
   }
   cppLanguageOption(): Cypress.Chainable {
     return cy.get("[data-cy='c++-value']");
