+++ packages/app/src/app/race/(play)/practice/page.tsx
@@ -13,7 +13,6 @@ import { RacePracticeCard } from "../../_components/practice/race-practice-card"
 import { Heading } from "@/components/ui/heading";
 import { languageTypes } from "@/lib/validations/room";
 
-
 type PracticeRacePageProps = {
   searchParams: {
     snippetId: string;
@@ -32,7 +31,9 @@ export default async function PracticeRacePage({
     snippet = await getSnippetById(searchParams.snippetId);
   } else if (searchParams.lang) {
     const language = searchParams.lang;
-    snippet = await getRandomSnippet({ language: languageTypes.parse(language) });
+    snippet = await getRandomSnippet({
+      language: languageTypes.parse(language),
+    });
   }
 
   return (
@@ -68,7 +69,8 @@ export default async function PracticeRacePage({
             </div>
             <Link
               href={
-                "/add-snippet?" + (searchParams.lang
+                "/add-snippet?" +
+                (searchParams.lang
                   ? `lang=${encodeURIComponent(searchParams.lang)}`
                   : `snippetId=${encodeURIComponent(searchParams.snippetId)}`)
               }
@@ -87,26 +89,44 @@ export default async function PracticeRacePage({
               description="How this game works and its caveats."
             />
             <ol className="list-decimal mt-4 pl-4 flex flex-col gap-4">
-              <li>The timer will start once you start typing on the displayed code you see above.</li>
-              <li>Once you mistype a character, you must fix it before you can continue typing.
+              <li>
+                The timer will start once you start typing on the displayed code
+                you see above.
               </li>
-              <li>You can start typing by clicking on the displayed code above.</li>
-              <li>The race will finish automatically once you complete the code snippet above.</li>
-              <li>If you are logged in, your data will be saved on the server for future references, such as:
+              <li>
+                Once you mistype a character, you must fix it before you can
+                continue typing.
+              </li>
+              <li>
+                You can start typing by clicking on the displayed code above.
+              </li>
+              <li>
+                The race will finish automatically once you complete the code
+                snippet above.
+              </li>
+              <li>
+                If you are logged in, your data will be saved on the server for
+                future references, such as:
                 <ol className="list-disc pl-8 mt-1 flex flex-col gap-2">
                   <li>Your Average CPM</li>
                   <li>Your Most Used Code Languages</li>
                 </ol>
               </li>
-              <li>When you press Enter without the ⏎ symbol and you are not on a whitespace, nothing will happen.</li>
-              <li>When you press Enter on a ⏎ symbol &#40;Going to a new line&#41;, 
-                you will be able to go to the next line and the whitespaces will
-                automatically be typed in for you.
+              <li>
+                When you press Enter without the ⏎ symbol and you are not on a
+                whitespace, nothing will happen.
+              </li>
+              <li>
+                When you press Enter on a ⏎ symbol &#40;Going to a new
+                line&#41;, you will be able to go to the next line and the
+                whitespaces will automatically be typed in for you.
               </li>
               <li>You can restart since this is practice mode.</li>
-              <li>If you click on get a new snippet and it&apos;s still the same, then there is not enough snippet for that language.
-                Please try creating a new one or keep hitting the get new snippet button.
-                </li>
+              <li>
+                If you click on get a new snippet and it&apos;s still the same,
+                then there is not enough snippet for that language. Please try
+                creating a new one or keep hitting the get new snippet button.
+              </li>
             </ol>
           </div>
         )}
