+++ packages/app/src/app/race/_components/common/progress-tracker.tsx
@@ -7,35 +7,35 @@ type Props = {
   name: string;
   image: string;
   progress: number;
-}
+};
 
-const ProgressTracker: React.FC<Props> = React.memo(({
-  name,
-  image,
-  progress
-}) => {
-
-  return (
-    <div className="relative flex items-center mb-5">
-      <ProgressBar>
-        <ProgressIndicator progress={progress} className="ease-in duration-200">
-          <Avatar 
-            style={{ left: (progress * 0.96) + "%" }}
-            className="w-7 h-7 absolute left-0 top-[-50%] border-2 border-monochrome rounded-full transition-[left] duration-200 ease-in"
+const ProgressTracker: React.FC<Props> = React.memo(
+  ({ name, image, progress }) => {
+    return (
+      <div className="relative flex items-center mb-5">
+        <ProgressBar>
+          <ProgressIndicator
+            progress={progress}
+            className="ease-in duration-200"
           >
-            <AvatarFallback>{name.charAt(0)}</AvatarFallback>
-            <AvatarImage
-              src={image}
-              alt={`${name}'s avatar`}
-              height={28}
-              width={28}
-            />
-          </Avatar>
-        </ProgressIndicator>
-      </ProgressBar>
-    </div>
-  );
-});
+            <Avatar
+              style={{ left: progress * 0.96 + "%" }}
+              className="w-7 h-7 absolute left-0 top-[-50%] border-2 border-monochrome rounded-full transition-[left] duration-200 ease-in"
+            >
+              <AvatarFallback>{name.charAt(0)}</AvatarFallback>
+              <AvatarImage
+                src={image}
+                alt={`${name}'s avatar`}
+                height={28}
+                width={28}
+              />
+            </Avatar>
+          </ProgressIndicator>
+        </ProgressBar>
+      </div>
+    );
+  }
+);
 
 ProgressTracker.displayName = "ProgressTracker";
